gestao_rh/urls.py
"""gestao_rh URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.1/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import include, path
from django.conf import settings
from django.conf.urls.static import static
from rest_framework import routers
from apps.core import views

from apps.funcionarios.api.view import FuncionarioViewSet
from apps.registro_hora_extra.api.view import RegistroHoraExtraViewSet

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'groups', views.GroupViewSet)
router.register(r'funcionarios', FuncionarioViewSet)
router.register(r'horas-extras', RegistroHoraExtraViewSet)


urlpatterns = [

    # REST
    path('api/', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),


    path("admin/", admin.site.urls),
    path('', include('apps.core.urls')),
    path('accounts/', include('django.contrib.auth.urls')),
    path('empresas/', include('apps.empresas.urls')),
    path('departamentos/', include('apps.departamentos.urls')),
    path('documentos/', include('apps.documentos.urls')),
    path('funcionarios/', include('apps.funcionarios.urls')),
    path('horas-extras/', include('apps.registro_hora_extra.urls')),
] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)  # para visualizar os arquivos (img, pdf, etc)
