requirements.txt
Django==4.1
djangorestframework
django-environ